requirements.txt
aiohttp==3.12.14
aiobotocore==2.23.1
gspread==6.2.1
pandas==2.3.1
pydantic==2.6.1
google-auth-oauthlib==1.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
fastapi==0.116.1
uvicorn==0.27.1 